Add In-Code Preview Support
