+++ manim/__main__.py
@@ -9,37 +9,8 @@ from manim.utils.module_ops import (
     get_scenes_to_render,
     scene_classes_from_file,
 )
-from manim.utils.file_ops import open_file as open_media_file
+from manim.utils.file_ops import open_file_if_needed
 from manim._config.main_utils import parse_args
-
-
-def open_file_if_needed(file_writer):
-    if config["verbosity"] != "DEBUG":
-        curr_stdout = sys.stdout
-        sys.stdout = open(os.devnull, "w")
-
-    open_file = any([config["preview"], config["show_in_file_browser"]])
-
-    if open_file:
-        file_paths = []
-
-        if config["save_last_frame"]:
-            file_paths.append(file_writer.image_file_path)
-        if config["write_to_movie"] and not config["save_as_gif"]:
-            file_paths.append(file_writer.movie_file_path)
-        if config["save_as_gif"]:
-            file_paths.append(file_writer.gif_file_path)
-
-        for file_path in file_paths:
-            if config["show_in_file_browser"]:
-                open_media_file(file_path, True)
-            if config["preview"]:
-                open_media_file(file_path, False)
-
-    if config["verbosity"] != "DEBUG":
-        sys.stdout.close()
-        sys.stdout = curr_stdout
-
 
 def main():
     console.print(f"Manim Community [green]v{__version__}[/green]")
@@ -101,8 +72,7 @@ def main():
             for SceneClass in scene_classes_from_file(input_file):
                 try:
                     scene = SceneClass()
-                    scene.render()
-                    open_file_if_needed(scene.renderer.file_writer)
+                    scene.render(preview=True)
                 except Exception:
                     console.print_exception()
