+++ manim/utils/file_ops.py
@@ -13,6 +13,7 @@ import os
 import platform
 import time
 import subprocess as sp
+from manim import config, logger
 from pathlib import Path
 
 from manim import __version__
@@ -77,3 +78,21 @@ def open_file(file_path, in_browser=False):
             raise OSError("Unable to identify your operating system...")
         commands.append(file_path)
         sp.Popen(commands)
+
+def open_file_if_needed(file_writer):
+    file_paths = []
+
+    if config["save_last_frame"]:
+        file_paths.append(file_writer.image_file_path)
+    if config["write_to_movie"] and not config["save_as_gif"]:
+        file_paths.append(file_writer.movie_file_path)
+    if config["save_as_gif"]:
+        file_paths.append(file_writer.gif_file_path)
+
+    for file_path in file_paths:
+        if config["show_in_file_browser"]:
+            open_file(file_path, True)
+        if config["preview"]:
+            open_file(file_path, False)
+
+        logger.info(f"Previewed File at: {file_path}")
